+++ src/main/java/com/woosan/hr_system/config/SecurityConfig.java
@@ -86,7 +86,7 @@ public class SecurityConfig {
                                 .frameOptions(HeadersConfigurer.FrameOptionsConfig::sameOrigin) // 동일 출처에서만 프레임 로딩 허용
                 )
 
-                // Authentication Entry Point 설정
+                // Authentication Entry Point 설정 - 401
                 .exceptionHandling(exceptionHandling -> exceptionHandling
                         .authenticationEntryPoint(customAuthenticationEntryPoint)
                 );
