+++ src/main/java/com/woosan/hr_system/employee/controller/EmployeeApiController.java
@@ -137,20 +137,6 @@ public class EmployeeApiController {
         return null;
     }
 
-    // 사원 영구 삭제
-    @RequireHRPermission
-    @DeleteMapping("/delete/{employeeId}")
-    public ResponseEntity<String> deleteEmployee(@PathVariable("employeeId") String employeeId) {
-        String message = employeeService.deleteEmployee(employeeId);
-        return switch (message) {
-            case "success" -> ResponseEntity.ok("사원이 삭제되었습니다.");
-            case "null" -> ResponseEntity.status(HttpStatus.NOT_FOUND).body("해당 사원을 찾을 수 없습니다.");
-            case "no_resignation" -> ResponseEntity.status(HttpStatus.BAD_REQUEST).body("해당 사원의 퇴사 정보가 없습니다.");
-            case "not_expired" -> ResponseEntity.status(HttpStatus.BAD_REQUEST).body("퇴사 후 1년이 지나지 않았습니다.");
-            default -> ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body("삭제하는 중 오류가 발생했습니다.");
-        };
-    }
-
     // 계정 잠금 설정
     @RequireHRPermission
     @PatchMapping("/set/accountLock/{employeeId}")
